services/messageservice/requirements.txt
-i https://pypi.org/simple
anyio==4.10.0; python_version >= '3.9'
asgiref==3.9.1; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
autobahn==24.4.2; python_version >= '3.9'
automat==25.4.16; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
cffi==2.0.0; platform_python_implementation != 'PyPy'
channels[daphne]==4.3.1; python_version >= '3.9'
channels-redis==4.3.0; python_version >= '3.9'
constantly==23.10.4; python_version >= '3.8'
cryptography==45.0.7; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.2.1
django==5.2.6; python_version >= '3.10'
djangorestframework==3.16.1; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
hyperlink==21.0.0
idna==3.10; python_version >= '3.6'
incremental==24.7.2; python_version >= '3.8'
msgpack==1.1.1; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; implementation_name != 'PyPy'
pyopenssl==25.2.0
redis==6.4.0; python_version >= '3.9'
service-identity==24.2.0
setuptools==80.9.0; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
twisted[tls]==25.5.0; python_full_version >= '3.8.0'
txaio==25.6.1; python_version >= '3.10'
typing-extensions==4.15.0; python_version >= '3.9'
zope-interface==8.0; python_version >= '3.9'
django-environ
django-cors-headers
djangorestframework-simplejwt
requests
httpx
django-prometheus==2.4.1; python_version >= '3.10'
